fix(): CleanCode

!171 merge clean-code into master

fix(): CleanCode

Created-by: fuyunde
Commit-by: fuyunde;weixin_43960572;shenjiahui_;liuhaoyu35;fanzhongheng;xiaocui235;zuyu;Archerls;ethWeek1;zhongyu1;lll0802;zd_gitcode;shirokawa_;m0_50464975;unknown
Merged-by: cann-robot
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR不涉及Issue，可填写"NA"。-->

<!--描述进行了哪些测试来验证你的改动。包括但不限于构造对应xx测试用例、二级冒烟、算子泛化等。-->

<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：

See merge request: cann/hccl!171

+++ build.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 # -----------------------------------------------------------------------------------------------------------
 # Copyright (c) 2025 Huawei Technologies Co., Ltd.
 # This program is free software, you can redistribute it and/or modify it under the terms and conditions of
@@ -109,7 +110,7 @@ function build_test() {
         LIBRARY_DIR="${LIBRARY_DIR}${ASCEND_HOME_PATH}/../../latest/opensdk/opensdk/gtest_shared/lib64:"
     fi
 
-    GCC_MAJOR=`gcc -dumpversion | cut -d. -f1`
+    GCC_MAJOR=$(gcc -dumpversion | cut -d. -f1)
     if [ "${ASAN}" == "true" ];then
         ARCH=$(uname -m)
         if [[ $ARCH == "x86_64" || $ARCH == "i386" || $ARCH == "i686" ]]; then
@@ -193,7 +194,7 @@ function build_ut() {
   local report_dir="${OUTPUT_PATH}/report/ut" && mk_dir "${report_dir}"
   cd "${BUILD_DIR}"
 
-  local LLT_KILL_TIME=1200
+  local llt_kill_time=1200
   CMAKE_ARGS="-DPRODUCT_SIDE=host \
               -DCMAKE_BUILD_TYPE=${CMAKE_BUILD_TYPE} \
               -DCMAKE_INSTALL_PREFIX=${OUTPUT_DIR} \
@@ -204,7 +205,7 @@ function build_ut() {
               -DENABLE_UT=${ENABLE_UT} \
               -DENABLE_ST=${ENABLE_ST} \
               -DOUTPUT_PATH=${OUTPUT_PATH} \
- 	          -DLLT_KILL_TIME=${LLT_KILL_TIME}"
+ 	          -DLLT_KILL_TIME=${llt_kill_time}"
 
   echo "CMAKE_ARGS=${CMAKE_ARGS}"
   cmake ${CMAKE_ARGS} ..
@@ -505,7 +506,7 @@ elif [ "${FULL_MODE}" == "true" ]; then
     cd .. & cd ${BUILD_DIR}
     CUSTOM_OPTION="${CURRENT_CUSTOM_OPTION} -DDEVICE_MODE=OFF"
     build_package
-    rm -rf ${BUILD_DEVICE_DIR}
+    [ -n "${BUILD_DEVICE_DIR}" ] && rm -rf ${BUILD_DEVICE_DIR}
 else
     CUSTOM_OPTION="${CUSTOM_OPTION} -DDEVICE_MODE=OFF"
     build_package
